copy_from_pqclean commit db0d5800c5c8a32df63414297671cd418ddc3129

+++ src/kem/kyber/pqclean_kyber512-90s_clean/reduce.c
@@ -18,7 +18,7 @@ int16_t PQCLEAN_KYBER51290S_CLEAN_montgomery_reduce(int32_t a) {
     int32_t t;
     int16_t u;
 
-    u = (int16_t)(a * QINV);
+    u = (int16_t)(a * (int64_t)QINV);
     t = (int32_t)u * KYBER_Q;
     t = a - t;
     t >>= 16;
